+++ Learn_Game/Arkanoid_Ball.h
@@ -8,34 +8,34 @@ namespace Arkanoid
 {
 	class Ball
 	{
-		float speed;
+		double speed;
 		sf::Vector2f position;
 		sf::Vector2f speedValue;
 		sf::Vector2f oldPosition;
-		float oryginalRadius;
-		float radius;
-		float currentSpeed;
-		float angle;
+		double oryginalRadius;
+		double radius;
+		double currentSpeed;
+		double angle;
 		bool updateAble;
-		void updateBuffs(const float dt);
+		void updateBuffs(const double dt);
 	public:
 		Ball();
-		Ball(sf::Vector2f position, float radius);
+		Ball(sf::Vector2f position, double radius);
 		virtual ~Ball()=default;
-		float speedMultipler;
-		void update(const float dt);
+		double speedMultipler;
+		void update(const double dt);
 		void increaseSpeed();
 		void decreaseSpeed();
 		void resetSpeed();
 		const sf::Vector2f getPosition() const;
-		const float getRadius() const;
-		void changeSize(float radius);
+		const double getRadius() const;
+		void changeSize(double radius);
 		void resetRadius();
-		void setRadius(float radius);
+		void setRadius(double radius);
 		void setHitAngle(const Arkanoid::Paddle & paddle);
 		void reverseXSpeed();
 		void reverseYSpeed();
-		const float getAngle() const;
+		const double getAngle() const;
 		void setPosition(sf::Vector2f pos);
 		void setUpdateAble(bool uA);
 		bool getUpdateAble();
