+++ Learn_Game/Arkanoid_SmallPaddleBuff.cpp
@@ -1,6 +1,6 @@
 #include "Arkanoid_SmallPaddleBuff.h"
 #include "Arkanoid_BuffFactory.h"
-Arkanoid::Buffs::SmallPaddleBuff::SmallPaddleBuff(float duration, Arkanoid::Paddle & paddle)
+Arkanoid::Buffs::SmallPaddleBuff::SmallPaddleBuff(double duration, Arkanoid::Paddle & paddle)
 	:Arkanoid::Buffs::IBuff(duration), paddlePTR(&paddle)
 {
 	goodEvilBuff = BuffType::EVIL_BUFF;
@@ -8,7 +8,7 @@ Arkanoid::Buffs::SmallPaddleBuff::SmallPaddleBuff(float duration, Arkanoid::Padd
 	paddlePTR->setSize(sf::Vector2f(paddlePTR->getSize().x / 2, paddlePTR->getSize().y));
 }
 
-void Arkanoid::Buffs::SmallPaddleBuff::update(const float dt)
+void Arkanoid::Buffs::SmallPaddleBuff::update(const double dt)
 {
 	removeLife(dt);
 }
